target-rootfs: limit number of files stripped at once
If a too large number of files is handed to strip, then this might lead
to "too many open files" errors on systems with restricted settings, in
turn causing a much bigger rootfs. Limit the number of files to be
stripped at once to 128, which should not hurt performance too badly.